close() -> utility
